Jaden_Menu-Functionality
I made all the sub-menus panels and made the pause and main menus functional*. I made a prefab for a button, but I don't want to implement it and brick the UI again. I have another project I need to work on now, but I'll update if I implement the button prefabs before next meeting.